add scripts to download PCAN router source code and hex2bin
download_pcan_source.sh:

The PCAN router source code is extracted from the "Development package
for ARM-based products" from the PEAK download site.

As only the source code of the different PCAN router hardware is needed
the unused non-Linux stuff is omitted in the unzip process and the PCAN
router code for the different hardware types is provided in the 'hw'
directory.

download_hex2bin109_source.sh:

Download and build the hex2bin tool from its original download site.
The Hex2bin Version 1.0.9 is also needed in the original "Development
package for ARM-based products" from the PEAK download site as this
version is limited to 16MB address space analogue to the pcanflash
protocol.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>